Switched from /star to /stardev. Use R1a recpars.
Also retain noise and rms files, and the 1024-pixel rimg previews.

It appears I was having a bad day for my initial commits storing
retrospectively the chronology for cohrs_run, supplying incorrect
commit dates for two of the three earlier versions.  The original
should read 2012-09-26 not a month later and Commit e1e152303e2e9
should be 2012-10-04 not days earlier.