scripts: Use an environment variable to make tiling lists more general.
These lists of CUBE_REBIN NDFs are used to form large mosaics up to 10
degrees in longitude from which the equally sized portions are
extracted for publication.  They are used in commands like that below.

  % wcsmosaic in=^in_inner1.txt out=$COHRS_TILED/tempcube_inner1 \
              method=gauss params="[0,1]" variance genvar conserve \
              accept

The environment variable COHRS_REDUCED must be defined and point to
the directory containing the reduced COHRS_<l>_<b>_CUBE_REBIN NDFs.