scripts: Add two mosaic sections and trim in velocity for equal-sized tile extraction.
The increased dimensions of the mosaics, in part because of weaves
rotated with respect to galactic latitude and longitude, caused HDS to
complain about the number (32-bit) of elements.  So broke the tiling
into two more sections (inner3 and middle2), and selected only the
velocity range where emission is present.  The tiles extents are as
follows.  Each has a 0.5-degree overlap at each end (except only 9.75
for the lower inner1 section).

inner1   10   18
inner2   18   25
inner3   25   32
middle1  32   39
middle2  39   46
outer1   46   54
outer2   54   62