filelists: Introduce new lists of files to combine into each 5-degree-wide mosaic.
The "_R2" in the names denotes Release 2.  The "n" suffix after the
"R2" arose because the original ??mosaic_R2.txt files created some
mosaics that contained too many elements for 32-bit addressing despite
a narrower width for the mosaics compared with Release 1.  That was
due to some outlying elements from k-mirror excursions and the wrong
co-ordinate system being chosen in the observation MSBs. The extra n
prevades through some scripts and msoaics, so it has stuck.

The need to change the mosaic ranges is because the tessarae files
have a wider spectral range to encompass the baseline either side of
the emission and have a higher resolution.  The consistent numerical
names make it easier to know where a given section should reside,
unlike than the former arbitrary "inner", "middle", and "outer"
with ordinal suffices.