recpars: Explictly control whether or not to align side bands.
This applies to multi-night reductions.

There is only one reduction that does have both USB and LSB
observations, hence ALIGN_SIDE_BAND is set to 1 for it.

Although Recpar ALIGN_SIDE_BAND defaults to off, there is good reason
to include it the recpar files.  For a while, the ORAC-DR primitive
_VERIFY_DATA_ had this defaulting to enable the alignment of sidebands
to cope with the aforementioed case.  However, this had a side effect
of mis-aligning uulti-night data taken in the same sideband.  As the
default in _VERIFY_DATA_ has already changed, it's better to assign
ALIGN_SIDE_BAND explicitly for multi-night reductions.