recpars: Introduce REF_SPECTRUM_FILE for three sections.
Around l=35.2 to 35.7 there is a sheet of emission at the same
velcoity as an off-position line.  The manual range that encompasses
that velocity, set via the REF_SPECTRUM_REGIONS recpar, is undermined
because the modal emission comes from the sheet not the reference
spectrum, and hence the reference line is not expunged.

To overcome this median spectra from the parts of the three sections
avoiding the emission were created, channels beyond the extent of the
reference line(s) set tot zero, and the recipe now adds this into the
reference spectra created in _SUBTRACT_REFERENCE_SPECTRUM_.  These
spectra were extracted with GAIA for l=35.7 to 36.0 to prove the
principle and then test the new ORAC-DR code.  Note that they could be
improved by masking the emission, say with an ARD polygonal section
and then running KAPPA:COLLAPSE with the mode estimator.