filelists: Pre-2016 data are in gzipped form.
While ORAC-DR can cope with the uncompresed files, in general
applications do not if the files are supplied via the ^ GRP facility.
So for example, it is not possible to see the velocity range of each
raw time-series file, because NDFTRACE will not access the gzipped
versions of the NDFs if the gal_*.list file names the uncompressed
NDFs.