status: Much improved log of reference-spectrum corrections.
It includes a key to the symbols in the Note column as well as column
titles.

There are numerous corrections to both original and filtered line
depths, the odd typo corrected, and some alignment tidying. [Note to
self: this file was called final_refspec_correction while retaining
the original refspec_correction.  It ought to have had numerous
commits... Now it will force a rewrite, so differences are best seen
in meld (or equivalent) comparing the two versions outside of git.]

Note this was compiled before improvements and additional correction
techniques were applied and further off-position signals identified by
comparing the median spectra in the overlaps between observed regions.