recpars: Set BASELINE_ORDER to 1 (V8 set).
In many regions the emission region spans a fractionally wide portion
of the spectral range, and using a higher order underestimates the
signal in the emission region, whereas a linear baseline does not. The
fourth-order fitting was originally chosen to cater for some
non-linear baselines and to match the order chosen for CHIMPS (Rigby
et al., 2015).  Comparisons with the likes of FUGIN (Umemoto et al,
2017) show a good match for integrated fluxes derived from the
spectral cubes made with linear-baseline subtraction, unlike those
constructed using fourth-order baseline subtraction.